task_11

+++ restaurants/views.py
@@ -2,33 +2,29 @@ from django.shortcuts import render, redirect
 from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
-
+def No(request):
+    return render(request, 'No.html')
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
         form = SignupForm(request.POST)
         if form.is_valid():
             user = form.save(commit=False)
-
             user.set_password(user.password)
             user.save()
-
             login(request, user)
             return redirect("restaurant-list")
     context = {
         "form":form,
     }
     return render(request, 'signup.html', context)
-
 def signin(request):
     form = SigninForm()
     if request.method == 'POST':
         form = SigninForm(request.POST)
         if form.is_valid():
-
             username = form.cleaned_data['username']
             password = form.cleaned_data['password']
-
             auth_user = authenticate(username=username, password=password)
             if auth_user is not None:
                 login(request, auth_user)
@@ -37,18 +33,14 @@ def signin(request):
         "form":form
     }
     return render(request, 'signin.html', context)
-
 def signout(request):
     logout(request)
     return redirect("signin")
-
 def restaurant_list(request):
     context = {
         "restaurants":Restaurant.objects.all()
     }
     return render(request, 'list.html', context)
-
-
 def restaurant_detail(request, restaurant_id):
     restaurant = Restaurant.objects.get(id=restaurant_id)
     items = Item.objects.filter(restaurant=restaurant)
@@ -57,9 +49,10 @@ def restaurant_detail(request, restaurant_id):
         "items": items,
     }
     return render(request, 'detail.html', context)
-
 def restaurant_create(request):
     form = RestaurantForm()
+    if request.user.is_anonymous:
+        return redirect('signin')
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
         if form.is_valid():
@@ -71,38 +64,44 @@ def restaurant_create(request):
         "form":form,
     }
     return render(request, 'create.html', context)
-
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
-    if request.method == "POST":
-        form = ItemForm(request.POST)
-        if form.is_valid():
-            item = form.save(commit=False)
-            item.restaurant = restaurant
-            item.save()
-            return redirect('restaurant-detail', restaurant_id)
-    context = {
-        "form":form,
-        "restaurant": restaurant,
-    }
-    return render(request, 'item_create.html', context)
-
+    if request.user.is_staff or request.user == restaurant.owner:
+        if request.method == "POST":
+            form = ItemForm(request.POST)
+            if form.is_valid():
+                item = form.save(commit=False)
+                item.restaurant = restaurant
+                item.save()
+                return redirect('restaurant-detail', restaurant_id)
+        context = {
+            "form":form,
+            "restaurant": restaurant,
+        }
+        return render(request, 'item_create.html', context)
+    else:
+        return redirect('No')
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    form = RestaurantForm(instance=restaurant_obj)
-    if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
-        if form.is_valid():
-            form.save()
-            return redirect('restaurant-list')
-    context = {
-        "restaurant_obj": restaurant_obj,
-        "form":form,
-    }
-    return render(request, 'update.html', context)
-
+    if request.user.is_staff or request.user == restaurant_obj.owner:
+        form = RestaurantForm(instance=restaurant_obj)
+        if request.method == "POST":
+            form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
+            if form.is_valid():
+                form.save()
+                return redirect('restaurant-list')
+        context = {
+            "restaurant_obj": restaurant_obj,
+            "form":form,
+        }
+        return render(request, 'update.html', context)
+    else:
+        return redirect('No')
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    restaurant_obj.delete()
+    if not request.user.is_staff:
+        return redirect('No')
+    else:
+        restaurant_obj.delete()
     return redirect('restaurant-list')
